Revert converting unsigned to signed & "fix" (not quite) calculations in conv.vhd
The data that we read from memory are pixel values, which are unsigned.
We need to perform signed calculations in conv.vhd, but the input and
output values are unsigned.

Regarding the calculations in conv.vhd, before they were wrong and
didn't compile. Now they are just wrong (but do compile)... I cannot
find the bug in the calculations, but it should be close.